Refactor models

+++ api/src/main/java/com/twitter/clone/api/repository/TweetsRepository.java
@@ -4,6 +4,6 @@ import com.twitter.clone.api.models.Tweet;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TweetsRepository extends JpaRepository<Tweet, Integer> {
+public interface TweetsRepository extends JpaRepository<Tweet, Long> {
 
 }
