+++ api/src/main/java/com/twitter/clone/api/repository/UsersRepository.java
@@ -4,6 +4,6 @@ import com.twitter.clone.api.models.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepository extends JpaRepository<User, Integer> {
+public interface UsersRepository extends JpaRepository<User, Long> {
 
 }
